feat: move ChatroomList FeedFilters to commons
Move ChatroomListKnownFeedFilter and ChatroomListNewFeedFilter from
amethyst module to commons module, enabling multiplatform reuse.

Changes:
- Add publicChatList, ephemeralChatList, isAllHidden to IAccount
- Add getOrCreatePublicChatChannel, getOrCreateEphemeralChannel to ICacheProvider
- Create ChatroomListKnownFeedFilter in commons (takes IAccount + ICacheProvider)
- Create ChatroomListNewFeedFilter in commons (takes IAccount)
- Replace amethyst originals with typealiases
- Update Account and LocalCache with override implementations
