+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/chats/rooms/dal/ChatroomListNewFeedFilter.kt
@@ -20,118 +20,4 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.chats.rooms.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.service.replace
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.quartz.nip17Dm.base.ChatroomKey
-import com.vitorpamplona.quartz.nip17Dm.base.ChatroomKeyable
-
-class ChatroomListNewFeedFilter(
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex
-
-    // returns the last Note of each user.
-    override fun feed(): List<Note> {
-        val chatList = account.chatroomList
-        val followingKeySet = account.followingKeySet()
-
-        val privateMessages =
-            chatList.rooms.mapNotNull { key, chatroom ->
-                if (!chatroom.senderIntersects(followingKeySet) && !chatList.hasSentMessagesTo(key) && !account.isAllHidden(key.users)) {
-                    chatroom.newestMessage
-                } else {
-                    null
-                }
-            }
-
-        return privateMessages.sortedWith(DefaultFeedOrder)
-    }
-
-    override fun updateListWith(
-        oldList: List<Note>,
-        newItems: Set<Note>,
-    ): List<Note> {
-        val me = account.userProfile()
-
-        // Gets the latest message by room from the new items.
-        val newRelevantPrivateMessages = filterRelevantPrivateMessages(newItems, account)
-
-        if (newRelevantPrivateMessages.isEmpty()) {
-            return oldList
-        }
-
-        var myNewList = oldList
-
-        newRelevantPrivateMessages.forEach { newNotePair ->
-            var hasUpdated = false
-            oldList.forEach { oldNote ->
-                val oldRoom = (oldNote.event as? ChatroomKeyable)?.chatroomKey(me.pubkeyHex)
-
-                if (newNotePair.key == oldRoom) {
-                    hasUpdated = true
-                    if ((newNotePair.value.createdAt() ?: 0L) > (oldNote.createdAt() ?: 0L)) {
-                        myNewList = myNewList.replace(oldNote, newNotePair.value)
-                    }
-                }
-            }
-            if (!hasUpdated) {
-                myNewList = myNewList.plus(newNotePair.value)
-            }
-        }
-
-        return sort(myNewList.toSet()).take(1000)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> {
-        // Gets the latest message by room from the new items.
-        val newRelevantPrivateMessages = filterRelevantPrivateMessages(newItems, account)
-
-        return if (newRelevantPrivateMessages.isEmpty()) {
-            emptySet()
-        } else {
-            newRelevantPrivateMessages.values.toSet()
-        }
-    }
-
-    private fun filterRelevantPrivateMessages(
-        newItems: Set<Note>,
-        account: Account,
-    ): MutableMap<ChatroomKey, Note> {
-        val me = account.userProfile()
-        val followingKeySet = account.followingKeySet()
-
-        val newRelevantPrivateMessages = mutableMapOf<ChatroomKey, Note>()
-        newItems.forEach { newNote ->
-            val noteEvent = newNote.event
-            if (noteEvent is ChatroomKeyable) {
-                val roomKey = noteEvent.chatroomKey(me.pubkeyHex)
-                val room = account.chatroomList.rooms.get(roomKey)
-
-                if (room != null &&
-                    (
-                        newNote.author?.pubkeyHex != me.pubkeyHex &&
-                            room.senderIntersects(followingKeySet) &&
-                            !account.chatroomList.hasSentMessagesTo(roomKey)
-                    ) &&
-                    !account.isAllHidden(roomKey.users)
-                ) {
-                    val lastNote = newRelevantPrivateMessages.get(roomKey)
-                    if (lastNote != null) {
-                        if ((newNote.createdAt() ?: 0L) > (lastNote.createdAt() ?: 0L)) {
-                            newRelevantPrivateMessages.put(roomKey, newNote)
-                        }
-                    } else {
-                        newRelevantPrivateMessages.put(roomKey, newNote)
-                    }
-                }
-            }
-        }
-
-        return newRelevantPrivateMessages
-    }
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+typealias ChatroomListNewFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.ChatroomListNewFeedFilter
