+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/model/IAccount.kt
@@ -20,8 +20,11 @@
  */
 package com.vitorpamplona.amethyst.commons.model
 
+import com.vitorpamplona.amethyst.commons.model.emphChat.EphemeralChatListState
 import com.vitorpamplona.amethyst.commons.model.marmotGroups.MarmotGroupList
+import com.vitorpamplona.amethyst.commons.model.nip28PublicChats.PublicChatListState
 import com.vitorpamplona.amethyst.commons.model.privateChats.ChatroomList
+import com.vitorpamplona.quartz.nip01Core.core.HexKey
 import com.vitorpamplona.quartz.nip01Core.signers.EventTemplate
 import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
 import com.vitorpamplona.quartz.nip04Dm.messages.PrivateDmEvent
@@ -104,6 +107,15 @@ interface IAccount {
 
     /** Marmot MLS group chat list */
     val marmotGroupList: MarmotGroupList
+
+    /** Public chat (NIP-28) list state */
+    val publicChatList: PublicChatListState
+
+    /** Ephemeral chat list state */
+    val ephemeralChatList: EphemeralChatListState
+
+    /** Check if all users in the set are hidden */
+    fun isAllHidden(users: Set<HexKey>): Boolean
 
     /** Whether a note is acceptable (not hidden, not blocked, etc.) */
     fun isAcceptable(note: Note): Boolean
