+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/feeds/ChatroomListNewFeedFilter.kt
@@ -1,0 +1,135 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.ui.feeds
+
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.Note
+import com.vitorpamplona.amethyst.commons.util.replace
+import com.vitorpamplona.quartz.nip17Dm.base.ChatroomKey
+import com.vitorpamplona.quartz.nip17Dm.base.ChatroomKeyable
+
+class ChatroomListNewFeedFilter(
+    val account: IAccount,
+) : AdditiveFeedFilter<Note>() {
+    override fun feedKey(): String = account.userProfile().pubkeyHex
+
+    // returns the last Note of each user.
+    override fun feed(): List<Note> {
+        val chatList = account.chatroomList
+        val followingKeySet = account.followingKeySet()
+
+        val privateMessages =
+            chatList.rooms.mapNotNull { key, chatroom ->
+                if (!chatroom.senderIntersects(followingKeySet) && !chatList.hasSentMessagesTo(key) && !account.isAllHidden(key.users)) {
+                    chatroom.newestMessage
+                } else {
+                    null
+                }
+            }
+
+        return privateMessages.sortedWith(DefaultFeedOrder)
+    }
+
+    override fun updateListWith(
+        oldList: List<Note>,
+        newItems: Set<Note>,
+    ): List<Note> {
+        val me = account.userProfile()
+
+        // Gets the latest message by room from the new items.
+        val newRelevantPrivateMessages = filterRelevantPrivateMessages(newItems, account)
+
+        if (newRelevantPrivateMessages.isEmpty()) {
+            return oldList
+        }
+
+        var myNewList = oldList
+
+        newRelevantPrivateMessages.forEach { newNotePair ->
+            var hasUpdated = false
+            oldList.forEach { oldNote ->
+                val oldRoom = (oldNote.event as? ChatroomKeyable)?.chatroomKey(me.pubkeyHex)
+
+                if (newNotePair.key == oldRoom) {
+                    hasUpdated = true
+                    if ((newNotePair.value.createdAt() ?: 0L) > (oldNote.createdAt() ?: 0L)) {
+                        myNewList = myNewList.replace(oldNote, newNotePair.value)
+                    }
+                }
+            }
+            if (!hasUpdated) {
+                myNewList = myNewList.plus(newNotePair.value)
+            }
+        }
+
+        return sort(myNewList.toSet()).take(1000)
+    }
+
+    override fun applyFilter(newItems: Set<Note>): Set<Note> {
+        // Gets the latest message by room from the new items.
+        val newRelevantPrivateMessages = filterRelevantPrivateMessages(newItems, account)
+
+        return if (newRelevantPrivateMessages.isEmpty()) {
+            emptySet()
+        } else {
+            newRelevantPrivateMessages.values.toSet()
+        }
+    }
+
+    private fun filterRelevantPrivateMessages(
+        newItems: Set<Note>,
+        account: IAccount,
+    ): MutableMap<ChatroomKey, Note> {
+        val me = account.userProfile()
+        val followingKeySet = account.followingKeySet()
+
+        val newRelevantPrivateMessages = mutableMapOf<ChatroomKey, Note>()
+        newItems.forEach { newNote ->
+            val noteEvent = newNote.event
+            if (noteEvent is ChatroomKeyable) {
+                val roomKey = noteEvent.chatroomKey(me.pubkeyHex)
+                val room = account.chatroomList.rooms.get(roomKey)
+
+                if (room != null &&
+                    (
+                        newNote.author?.pubkeyHex != me.pubkeyHex &&
+                            room.senderIntersects(followingKeySet) &&
+                            !account.chatroomList.hasSentMessagesTo(roomKey)
+                    ) &&
+                    !account.isAllHidden(roomKey.users)
+                ) {
+                    val lastNote = newRelevantPrivateMessages.get(roomKey)
+                    if (lastNote != null) {
+                        if ((newNote.createdAt() ?: 0L) > (lastNote.createdAt() ?: 0L)) {
+                            newRelevantPrivateMessages.put(roomKey, newNote)
+                        }
+                    } else {
+                        newRelevantPrivateMessages.put(roomKey, newNote)
+                    }
+                }
+            }
+        }
+
+        return newRelevantPrivateMessages
+    }
+
+    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
+}
